cleaning up code, commenting and organinzing into separate files base on functionality
